requirements.txt
# Scrapy core
scrapy==2.11.2

# Cloudflare bypass
curl-cffi==0.7.4
playwright==1.45.0

# HTML parsing
parsel==1.9.1
lxml==5.2.2

# Utilities
python-dotenv==1.0.1
dateparser
psycopg2-binary
